+++ packages/akashic-cli-serve/src/client/store/storage.ts
@@ -5,7 +5,6 @@ export interface StorageData {
 	playerId: string;
 	playerName: string;
 	showsDevtools: boolean;
-	showsBgImage: boolean;
 	devtoolsHeight: number;
 	activeDevtool: string;
 	showsEventList: boolean;
@@ -15,6 +14,8 @@ export interface StorageData {
 	instanceArgumentListWidth: number;
 	instanceArgumentEditContent: string;
 	joinsAutomatically: boolean;
+	showsBackgroundImage: boolean;
+	showsGrid: boolean;
 }
 
 function asBool(s: string | null): boolean | null {
@@ -78,7 +79,9 @@ export class Storage {
 			selectedArgumentName: choose(getQueryValue(qp.selectedArgumentName), s.selectedArgumentName, null),
 			instanceArgumentListWidth: choose(asNumber(getQueryValue(qp.instanceArgumentListWidth)), s.instanceArgumentListWidth, 150),
 			instanceArgumentEditContent: choose(getQueryValue(qp.instanceArgumentEditContent), s.instanceArgumentEditContent, ""),
-			joinsAutomatically: choose(asBool(getQueryValue(qp.joinsAutomatically)), s.joinsAutomatically, false)
+			joinsAutomatically: choose(asBool(getQueryValue(qp.joinsAutomatically)), s.joinsAutomatically, false),
+			showsBackgroundImage: choose(asBool(getQueryValue(qp.showsBackgroundImage)), s.showsBackgroundImage, false),
+			showsGrid: choose(asBool(getQueryValue(qp.showsGrid)), s.showsGrid, false)
 		});
 	}