+++ packages/akashic-cli-serve/src/client/view/atom/ToolLabel.tsx
@@ -10,9 +10,8 @@ export interface ToolLabelProps {
 @observer
 export class ToolLabel extends React.Component<ToolLabelProps, {}> {
 	render() {
-		const { title, children, emphasizeBorder} = this.props;
-		const borderStyle = emphasizeBorder ? "border-emphasize" : "";
-		return <p className={[styles["tool-label"], styles[borderStyle]].join(" ")} title={title}>
+		const { title, children } = this.props;
+		return <p className={styles["tool-label"]} title={title}>
 			{ children }
 		</p>;
 	}