+++ packages/akashic-cli-serve/src/client/view/container/ToolbarContainer.tsx
@@ -8,6 +8,7 @@ import { Operator } from "../../operator/Operator";
 import { PlayControlPropsData } from "../molecule/PlayControl";
 import { InstanceControlPropsData } from "../molecule/InstanceControl";
 import { PlayerControlPropsData } from "../molecule/PlayerControl";
+import { DisplayOptionControlPropsData } from "../molecule/DisplayOptionControl";
 import { ToolBar } from "../organism/ToolBar";
 
 export interface ToolBarContainerProps {
@@ -26,14 +27,13 @@ export class ToolBarContainer extends React.Component<ToolBarContainerProps, {}>
 			makePlayControlProps={this._makePlayControlProps}
 			makeInstanceControlProps={this._makeInstanceControlProps}
 			makePlayerControlProps={this._makePlayerControlProps}
+			makeDisplayOptionControlProps={this._makeDisplayOptionControlProps}
 			showsAppearance={toolBarUiStore.showsAppearanceMenu}
 			showsDevtools={toolBarUiStore.showsDevtools}
-			showsBgImage={toolBarUiStore.showsBgImage}
 			showsInstanceControl={(localInstance.executionMode === "replay") || toolBarUiStore.showsDevtools}
 			targetService={targetService}
 			onToggleAppearance={operator.ui.toggleShowAppearance}
 			onToggleDevTools={operator.ui.toggleShowDevtools}
-			onToggleBgImage={operator.ui.toggleShowBgImage}
 		/>;
 	}
 
@@ -74,6 +74,18 @@ export class ToolBarContainer extends React.Component<ToolBarContainerProps, {}>
 			isJoined: localInstance.isJoined,
 			isJoinEnabled: (localInstance.executionMode === "passive" && joinEnabled),
 			onClickJoinLeave: operator.play.toggleJoinLeaveSelf
+		};
+	}
+
+	private _makeDisplayOptionControlProps = (): DisplayOptionControlPropsData => {
+		const { operator, toolBarUiStore } = this.props;
+		return {
+			showsDisplayOptionPopover: toolBarUiStore.showsDisplayOptionPopover,
+			showsBackgroundImage: toolBarUiStore.showsBackgroundImage,
+			showsGrid: toolBarUiStore.showsGrid,
+			onToggleDisplayOptionPopover: operator.ui.toggleShowDisplayOptionPopover,
+			onToggleShowBackgroundImage: operator.ui.toggleShowBackgroundImage,
+			onToggleShowGrid: operator.ui.toggleShowGrid
 		};
 	}
 }